Add sample unit test for readme
